refactor: omit unnecessary logic branch
Previously the code listened for `close`. This was changed to `end` in
the previous commit since an input buffer may not have an fd to close.

We do not need to test for cancelation in the new `end` handler because
as per the yazul docs "After [the `error` event] has been emitted, no
further `entry`, `end`, or `error` events will be emitted."